#2674 Cleaned code duplication after moving logic into correct files.
