+++ backend/census_historical_migration/workbooklib/secondary_auditors.py
@@ -1,4 +1,5 @@
 from census_historical_migration.workbooklib.excel_creation_utils import (
+    add_hyphen_to_zip,
     set_uei,
     map_simple_columns,
     generate_dissemination_test_table,
@@ -6,7 +7,6 @@ from census_historical_migration.workbooklib.excel_creation_utils import (
 from census_historical_migration.base_field_maps import SheetFieldMap
 from census_historical_migration.workbooklib.templates import sections_to_template_paths
 from census_historical_migration.workbooklib.census_models.census import dynamic_import
-from census_historical_migration.workbooklib.sac_creation import add_hyphen_to_zip
 from audit.fixtures.excel import FORM_SECTIONS
 
 import openpyxl as pyxl
