+++ backend/census_historical_migration/workbooklib/workbook_creation.py
@@ -1,12 +1,9 @@
+from django.core.files.uploadedfile import SimpleUploadedFile
 from fs.memoryfs import MemoryFS
 
 from census_historical_migration.workbooklib.sac_creation import (
     _post_upload_workbook,
-    _make_excel_file,
-    _create_test_sac,
 )
-
-from django.apps import apps
 
 from .utils import get_template_name_for_section
 
@@ -15,21 +12,11 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-def setup_sac(user, test_name, dbkey):
-    if user is None:
-        logger.error("No user provided to setup_sac")
-        return
-    logger.info(f"Creating a SAC object for {user}, {test_name}")
-    SingleAuditChecklist = apps.get_model("audit.SingleAuditChecklist")
-
-    sac = SingleAuditChecklist.objects.filter(
-        submitted_by=user, general_information__auditee_name=test_name
-    ).first()
-
-    logger.info(sac)
-    if sac is None:
-        sac = _create_test_sac(user, test_name, dbkey)
-    return sac
+def _make_excel_file(filename, f_obj):
+    content = f_obj.read()
+    f_obj.seek(0)
+    file = SimpleUploadedFile(filename, content, "application/vnd.ms-excel")
+    return file
 
 
 def generate_workbook(workbook_generator, dbkey, year, section):
